Improvements and bug fixes to SharedMemory and Semaphore
SharedMemory.php
. The SharedMemory implementation had a flaw where in the keyMapper instance variable which maps non-numeric keys for the array backed by shared memory to an integer which can be used by the Sys V shm implementation.  This variable is not stored in the shared memory segment, thus each process accessing the shared memory has its own copy.  This results in duplicates arising in the shared memory for the same key if different processes attempt to create the same key in the shared memory.

This keyMapping array is now stored in index 0 of the shared memory and thus shared by all processes accessing the shared memory.

. Added the ability to use the mutex instance inside the SharedMemory object externally, so that non-atomic operations on the shared memory (such as performing a lookup to see if a key exists before adding something for that key) can be performed without creating race conditions.  So now there is a lock method for SharedMemory which will lock the shared memory, multiple accesses can occur, and then the release method can be called which unlocks the shared memory.

To implement the above feature, changes were made to SharedMemory's internal locking mechanism with the mutex by checking first to see if the Semaphore has already acquired the lock (in this process) before attempting to lock it again.  Locking again would could deadlock.

. Added method to SharedMemory so that getKeys returns just the array keys of the SharedMemory object.  Provides same result as calling array_keys function on an array.

A new protected method was created to replace getKeys for internal use called getKeyMap.  getKeyMap returns the entire key=>value array as is used internally by SharedMemory.

Semaphore.php
Ensure that the lockExecute method actually updates the isAcquired instance variable during its call so that calls to this method by SharedMemory will know that it is already locked and not attempt to lock again.